requirements.txt
fastapi>=0.109.0
uvicorn>=0.25.0
SQLAlchemy>=2.0.0
pydantic>=2.0.0
pydantic-settings>=2.0.0
python-jose>=3.3.0
passlib>=1.7.4
python-dotenv>=1.0.0
python-multipart>=0.0.5
email-validator>=2.0.0
bcrypt>=4.0.0
alembic>=1.13.0
pytest>=7.4.0
httpx>=0.25.0 